+++ server/lib/python/cartodb_services/cartodb_services/config/legacy_rate_limits.py
