+++ server/lib/python/cartodb_services/cartodb_services/tools/service_manager.py
@@ -2,12 +2,8 @@ from cartodb_services.metrics import QuotaService
 from cartodb_services.tools import Logger
 from cartodb_services.tools import RateLimiter
 from cartodb_services.refactor.tools.logger import LoggerConfigBuilder
-from cartodb_services.refactor.core.environment import ServerEnvironmentBuilder
-from cartodb_services.refactor.backend.server_config import ServerConfigBackendFactory
-from cartodb_services.refactor.backend.user_config import UserConfigBackendFactory
-from cartodb_services.refactor.backend.org_config import OrgConfigBackendFactory
 from cartodb_services.refactor.backend.redis_metrics_connection import RedisMetricsConnectionFactory
-from cartodb_services.refactor.config import RateLimitsConfigBuilder
+from cartodb_services.config import ServiceConfiguration, RateLimitsConfigBuilder
 
 class RateLimitExceeded(Exception):
     def __str__(self):
@@ -52,31 +48,6 @@ class ServiceManagerBase:
     @property
     def logger(self):
         return self.logger
-
-
-class ServiceConfiguration:
-    def __init__(self, service, username, orgname):
-        self._server_config_backend = ServerConfigBackendFactory().get()
-        self._environment = ServerEnvironmentBuilder(server_config_backend).get()
-        self._user_config_backend = UserConfigBackendFactory(username, environment, server_config_backend).get()
-        self._org_config_backend = OrgConfigBackendFactory(orgname, environment, server_config_backend).get()
-
-    @property
-    def environment(self):
-        return self._environment
-
-    @property
-    def server(self):
-        return self._server_config_backend
-
-    @property
-    def user(self):
-        return self._user_config_backend
-
-    @property
-    def org(self):
-        return self._org_config_backend
-
 class ServiceManager(ServiceManagerBase):
     """
     This service manager delegates the configuration parameter details,
@@ -85,7 +56,7 @@ class ServiceManager(ServiceManagerBase):
     """
 
     def __init__(self, service, config_builder, username, orgname):
-        service_config = ServerConfiguration(service, username, orgname)
+        service_config = ServiceConfiguration(service, username, orgname)
 
         logger_config = LoggerConfigBuilder(service_config.environment, service_config.server).get()
         self.logger = Logger(logger_config)
